working ramp

+++ src/Robot/MotorControl.h
@@ -88,7 +88,7 @@ public:
   int Percent2RPM(float pct);
   float RPM2Percent(int rpm);
 
-  int ramp(float requestedPower, float accelRate);
+  float ramp(float requestedPower, float accelRate);
 };
 
 // MotorControl* GlobalClassPointer[MAX_NUM_ENCODERS];
